\name{PolytomousUtils-package}
\alias{PolytomousUtils-package}
\alias{PolytomousUtils}
\docType{package}
\title{
\packageTitle{PolytomousUtils}
}
\description{
\packageDescription{PolytomousUtils}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{PolytomousUtils}

Maintainer: \packageMaintainer{PolytomousUtils}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = PolytomousUtils)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{PolytomousUtils}}
%\section{Documentation Index}{\packageIndices{PolytomousUtils}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
